Re #17114. Fixed workspace factory size check that broke axis init.
This causes some system test failures. The problem is that the semantics
for "differentSize" passed to initializeFromParent are slightly
different. In the old-style factory this depended on whether Y was
specified or a default value, i.e., it would set differentSize=true even
if the size that was explicitly specified was identical to the parent
size. This seems like a quite unexpected behavior/side-effect.