Refs #17521 Fix for issue with MultiPeriod group determination
Previously the MultiPeriodGroups could be mixed in with standard workspaces. The standard workspaces were silently dropped, so it was only by chance that the test test_mixed_multiperiod_group_and_non_multiperiod_group_inputs_throws was passing. This test should now pass for the correct reasons, with a mix of workspace types throwing an error.