Update secondary axis after fit completed
Fixes a bug noticed when fitting muon data (though applies to any):
On fit completion, PPtool adds curves to graph. This alters left Y axis.
Right Y axis is not updated to match, and they become out of sync.
Add a call to update the secondary axis, if they should be in sync.

re #17770