Split interface into two interfaces
MuonAnalysisFitDataPresenter held an IWorkspaceFitControl* - this interface
was implemented by the base FitPropertyBrowser, meaning that some
muon-specific methods on the interface were not implemented in the base
class.

It is better to move the muon-specific methods to a new interface,
IMuonFitDataModel, that is implemented only by MuonFitPropertyBrowser and
not by the base FitPropertyBrowser.

re #17973