Prevent stale plot guesses being stuck on the graph
When multi fit mode is on, need to remove any guess before updating
function in the model, and then replot afterwards. This also updates the guess.

There is no need to do this if multi fit mode is off as there you are setting the
function directly in the model, and it works fine as is.

re #18060