Refs #18078 Fix LET system tests
Using SpectrumInfo to calculate the distance results in a small (9th s.f.) difference in the distance calulated to the source. This leads to an error in the LETReduction system tests. The tolerance of compare workspaces is increased to allow for this in LETReduction. In LETReductionEvent2015Multirep there are two bins with a greater discrepancy than 1e-5, new comparison files were created for this.