Refs #14793 Mask workspace with hardware grouping.
In some cases the hardware grouping can cause a workspace to have more
than on detector per spectrum. When we have a mask where each spectrum
corresponds to a detector and the workspace to be masked has a smaller
number of spectra we should append the indicies to the detectorList.