Refs #14793 Fix issue with workspace of detectors
This adds the ability to pass a workspace which is not a mask workspace
but a list of spectra corresponding to detectors to be masked. This is
useful in the case where the number of spectra do not match but the
number of detectors do, i.e. when some grouping has been applied to the
workspace.