Fix debug assertion when running tests in Windows debug mode.
The vector_iterator::+ was first passed the value of m_maxX and then the
result of this has m_histogram subtracted. m_maxX cause the iterator to
go out of bounds and the strict checking on MSVC throws a debug assertion.
The parentheses simply ensure the number is correct before passing to
iterator+.