Re #18219 Updating unit test acording to last changes
Note that some tests have been removed in ReflectometryReductionOne2Test. The reason is that those tests were testing the rebinned workspace in Q which no longer happens in this algorithm but in ReflectometryReductionOneAuto. Note that three tests were removed from ReflectometryReductionOneAuto2Test. The reason is that the algorithms wouldn't run anymore due to Rebin, which now has strict tests for rebin params. Rebin was failing in the removed tests because the input workspaces used were not appropriate for the reduction, for instance, the OFFSPEC was a scan run.