Ensure output EventLists have correct sort order after filtering
Previously the sort order was left untouched and this left a route for
events to be added without affecting the sort order. This would create
an inconsistent state if for example the sort order was set to TOF and
then filtering was applied. After filtering the list wuld still think it
was sorted by TOF and operations such as rebinning would give incorrect
results.
Refs #18477