Re #19001. Undo detector move in DiffractionEventCalibrateDetectors.
If we do not do this we get position updates in `detList`. The old
instrument code did maintain positions, even if the corresponding
instrument was updated, but now updates show up since the COW mechanism
in ExperimentInfo::instrumentParameters was removed.

This should also leave the `InputWorkspace` unmodified. Previously it
was modified on algorithm exit.