Re #19151. Remove detector grouping from IndexInfo constructor.
This would be weird in the MPI case, since spectrum numbers passed to
the constructor are a global list, but groupings are set only for the
local spectra. Potentially we even need filtering for the grouping list,
after IndexInfo has been constructed, so requiring it in the constructor
does not make sense.