Add CacheComponentVisitor
ComponentVisitor types, such as InfoComponentVisitor, build the component index list in a well
defined and specific way. InfoComponentVisitor cannot be properly used outside of an ExperimentInfo because
a ParameterMap is necessary in order to extract parameterised information such as positions and locations.
However, we need to be able to support ComponentID -> ComponentIndex type mappings when an ExperimentInfo
may not be present, such as directly on the Instrument. For example, Instrument::componentIndex(ComponentID).

CacheComponentVisitor solves this problem. It gurantees to give the same ComponentID -> ComponentIndex mapping that
the InfoComponentVisitor, and hence API::ComponentInfo will.

One particular problem is that Instrument1.0 badly  mixes construction with other aspects
f the instrument behaviour. For example, the instrument tree is never reasonablly fixed in creation.
i.e. you can call ::add any time you wish, and that can easily lead to an invalid state. For that reason, there is a
good_cache flag for the ComponentID map so that a rebuild can be identified.

Note that one could still bypass my safety mechanism, by say adding a bank to an instrument, and then using the reference on the client
side to add subcomponents.