Re #18829 Make output order of multiple groups consistent
When summing in Q, make the output order the same as when summing in
lambda. The latter uses GroupDetectors, which orders the groups by the
first spectrum number in the group. This commit sorts the detector
groups in the same order so it will be the same for any further
processing.