Re #18829 Add unit tests for sum in Q and change detector positions
- Previously the unit tests were using 3 detectors at the same position.
This commit changes them to be at different positions, which is required
for meaningful tests for the summation in Q functionality.
- Some test results have therefore changed as a result of the new
detector positions. To limit the changes, I have used processing
instructions of "2" where only a single detector is used. This is the
centre detector and has the same position as it did previously.
- The order of the processing instructions in the tests for multiple detector groups have
been modified to check that the resulting groups are being sorted correctly.
- Unit tests for summing in Q added.