Re #19773 Avoid hidden copy constructor
This commit adds a copy constructor that takes a non-const argument.
This is required because otherwise the templated constructor would take
precedence over the copy constructor when supplied with a non-const argument.

Currently the templated constructor is functionally the same as the new copy
constructor. However, adding the new copy constructor avoids ambiguity, and
means we can add some safety by casting the argument to const before passing
it to the base class constructor.