Re #19773 Fix copy-constructor warning
This commit removes a second copy constructor, which was causing a
compiler warning. The second constructor had been added to avoid a
problem where the copy constructor is hidden by a templated constructor
when the argument is non-const. The templated constructor has the
correct behaviour, though, so the second copy constructor is not
really necessary.