Add constrainAll to composite function re #19010
In the course of this, I had to make sure [] returns a CompositeFunctionWrapper when it is a composite function, so that it works for nested composite functions.

Also this was implemented by a crude try and do nothing in exception. This unfortunately does not silence c++ errors arising from the function not having a given parameter.

Signed-off-by: Karl Palmen <karl.palmen@stfc.ac.uk>