Re #19278. Adjusted deltas for unstable fit parameters in test.
There is changed code in, e.g., adjustUpSampleAndSourcePositions that I
thought initially was responsible for this. However, it seems that the
offsets passed to this function are changed, i.e., this seems unrelated
and might be due to a different computation order when computing
relative component positions, leading to different fit results?