Re #19278. Remove test of weird (buggy?) behavior in Crystal helper.
I think the older behavior of adjustUpSampleAndSourcePositions took the
position of the source in the instrument passed to it as the base
position, and somehow allowed the instrument and DetectorInfo (now
ComponentInfo) to be based on different workspaces. This does not seem
to make sense, and thus the test is also questionable. I cannot exclude
that I misunderstand the purpose of this though.