Fix incorrect cropping of input
When summing in Q, the input workspace was being cropped, causing counts
to be missing from the summed result. The crop is now done after the
summation. There was also a problem calculating the crop range because
this was being based on input wavelength. The crop is now done on the
projected wavelength range instead.

Re #20120