A quick patch for issue #20256
If there is anything in the ads that isn't a WorkspaceGroup or a MatrixWorkspace, such as a TableWorkspace, then has_hash throws an exception.  I'm making the argument that, if it's not a MatrixWorkspace, then it doesn't have the hash, so the answer is simply false and not an exception.  This fixes the issue #20256, since it's now possible to search through the ads for a hash without the exception crashing the routine.