refs #20264. Check for null shape
This is something that was done in Instrument 1.0 and as far as possible I have made backward compatible in the Geometry::ComponentInfo layer. Since shape returns a const ref, I've exposed the hasShape method publicly so that clients may check a shape before extracting without introducing uncessary slowdown by checking it as part of the shape() method.