refs #20264. Remove annoying ref args
Presumably there were initial concerns that RVO would not take place. As a result the API for the BoundingBox is strange, and I do not want to replicate that in the new API for ComponentInfo. One thing that tripped me up was the fact that the ObjComponent::getBoundingBox was reassigning the input arg, not calling ::grow, for example, there is also some mess around the coordinate system which is stripped from the input. Eventually I would like to remove the reference BoundingBox argument altogether, but for the moment I keep it for backwards compatibility.