Revert change to obtain height generically for all shapes
Whilst height is a common property for FlatPlate and Cylinder, they are not the same widget and therefore can not be accessed by one call.
It could be possible to access them via one widget, but I am reluctant to change widget names this close to a release.
Would be good to make this change after release.

Refs #20458