Temporarily fix the TOF axis.
The binning was off by one, and the phase of chopper 1 was read from a
wrong NeXus field. However, the correct field
'VirtualChopper.chopper1_average_phase' seems to contain the value of
'VirtualChopper.chopper2_average_speed' and vice versa. So at the
moment, we are reading the phase from the chopper 2 average speed field.

Re #20978