delete a (not empty) line
addComponent was used accidentally and causes an unhandled exception in ReflectometryReductionOneTest.
In general, adding a component twice should result in a better (error, warning) handling in Mantid. I could not debug easily. I created two workspaces and `CompareWorkspaces` resulted in true but one fulfilled the unit test and the other not!