Move non-ParaView plugins back to bin and exclude Qt5
Windows needs the DLLs and executables to be in the same directory
or it can't find the dependencies. This can be solved if running through
a wrapper script but the IDE calls th .exe directly and not allowing
this would degrade the developer experience alot.
This only affects the developer build and not the installed locations.
Refs #20288