Reorder code to reduce complexity and duplication
Previously, the processing algorithm defaults were being set after preprocessing was done. If the same property was being used for both preprocessing and processing, we had to remember which columns had been preprocessed and set as inputs to the processing algorithm so as not to overwrite them with the default algorithm properties. This commit changes this so that defaults are set first, and are overwritten by preprocessed values afterwards

Re #21218