WIP: collect properties in a map rather than setting on algorithm
We can't set non-preprocessed default options on the algorithm because they may need preprocessing before they are valid algorithm properties. This commit gathers them into a map first and sets them on the algorithm as the last step.

This needs further work to pass the correct values to preprocessing from the global options.

Re #21218