Avoid supplying workspace to total_scattering function
This is to match other isis_powder style functions (e.g. create_vanadium and focus)

Given the nature of it being post reduction analysis, it might be worth adding the ability to provide a workspace.
This would have to be discussed with an instrument scientist to see if it of use.

Refs #21294