Re #21097: Made RunFile event mode aware.
RunFinder now uses sans file_information behind the scenes to determine
the display name and whether or not a file contains event data.

RunFinder and RunFile are renamed to SummationRunFinder and
SummationRunFile to give them more of a destinction from a sans
file_information object.

These changes enable a correct implementation of has_any_event_data in the
RunSelection class.