Update GenerateNotebook to use canonical options
GenerateNotebook now uses the same functions for finding the options as GenericDataProcessorPresenter::reduceRow and should give the same results. Note that the output workspace names are still done in a different way and need consolidating as well. They need to be specified by OutputWorkspace properties because they contain special characters e.g. '+'

Re #21457