Convert PythonCodeExecution callbacks to Qt signals
It makes it easier to integrate with other code. This uncovered an
error in the code execution that has been fixed here. Previous changes
were not always executing the code as blocks and this caused
some tests to pass when they would fail in the block-based scenario.
We now only execute as blocks so the test now covers this.
Refs #21251