Re #21482 Add scripts to automatically install GSAS-II
A Python script to download some revision of GSAS-II and then install in
an arbitrary directory was written. Two .bat files were also implemented
to wrap around it, one of which downloads the latest revision, the other
downloads revision 3216 (known to work with GSASIIRefineFitPeaks). Both
algorithms by default install GSAS-II in C drive