Add the mantid3d projection for 3d plotting
Adds a new projection for accessing 3d plotting in matplotlib via the workbench.
This functionality allows users to supply a workspace rather than a array of data.
This change also covered a large amount of white space correction in both the init and functions file, to conform to the PEP8 style guide

Refs #21720