Re #20991: Update LoadDetectorInfoTest to use delta for comparison
LoadDetectorInfo appeared to be using castaround and substr to check
precisions of loaded values to n decimal places. This commit changes
to the use of TS_ASSERT_DELTA using 1.e-n to achieve the same
effect.

This also simplified the test cases and reduces the required
precision in some cases due to the fact that for detector positions 1.e-3 is
accurate to the milimetre anyway.