Re #21640 Make plotting to separate window use ws from model
Previously, if the ws was in the ADS, the version from the ADS would
be used. However the pointers don't point to the same workspace,
so plotToSeparateWindow would plot different things, depending on whether
the workspace exists in mtd or not. Now, the ws from the model is always used