Restructured solution to fix failing unit test.
The HYSA file contains event monitors which have no event_id attribute.
The check has now changed to ensure that iff the event_id key exists on
the monitor then it has more than one element. AKA if the event_id does
not exist but the other event keys do then the monitor is classified as
an event monitor.

Re #21818