Remove queues of rows/groups to process
This commit removes the explicit state machine which keeps track of which row/queue to process next. Instead, we use the state that is stored on per-row/group basis to determine which item is next to process. The latter needs to be done anyway so this commit simplifies the logic.

Re #17029