Fix a bug where rows can lose their processed state
After doing a transfer of new rows into the table, there was a brute force
reset of the model, which meant that the processed state of rows was being
lost. This meant that rows could be re-processed unnecessarily, and with the
other changes in this PR caused a bug where the workspace names could not be
found for plotting.

This commit removes the reset and instead adds the new rows to the existing
model. It also moves the logic for editing the underlying table into the model
so the tree manager no longer needs to do this.

Re #20088