Re #21921: Limited pallete patching to Unity.
Finding a general purpose fix for the tooltip colour is a longer
term problem as it depends on the chosen mantid theme in
View > Preferences.

Since the Adwaita Qt plugin which is installed by default on Fedora
works. This commit limits the fix to only apply to Unity Desktop
users where the adwaita-qt package is not installed by default and
none of the available themes work.

This means that RHEL users not previously affected by the issue are not
broken by the original PR to fix it.