refs #20266 Instrument for PMap must  be neutronic
The reasoning is that the ExperimentInfo sets the instrument on the
ParameterMap. We only create ExperimentInfos (MatrixWorkspace etc)
around neutronic instruments. By enforcing this (if it works) then we
can better reason that if a Component has a PMap, the held ComponentInfo
corresponds to the neutronic instrument. Physical Instruments are only
used in the InstrumentView, and that is drawn directly from
ComponentInfo, so no problems there.