Implemented row insertion notification.
This required making JobTreeView aware of the editor status.
When a presenter is notified of a row insertion it needs to be able
to remove the inserted row if the row is greater than the maximum
depth for example.

However after inserting a new row we natuarally open the editor at it
and set it as the current index. If you try to remove the row while
it is the current index and the editor is open then the row is
silently not removed.

Before attempting to remove the row we try detect if the removal would
be inefective and if so, shift the current index somewhere else, if t
here is nowhere else to move it then we thow an exception.

This commit therefore makes it impossible to delete all rows,
something we may need to fix in a future commit.

Re #22263