IndexProperty now valid if workspace pointer is null
It is assumed that this is the case of the empty set and should
be considered valid. The current mantid.simpleapi computes
an order for the function parameters based on the validity of the
properties. An empty set should be considered valid so that it is not
required to be supplied by a user and becomes an optional property.
Refs #22332