Add the ability to update existing changed rows
Previously, rows that had been changed were considered to be different to the original row, so that if you attempted to re-add the row, then a duplicate would be added because the rows do not match. This commit changes this to identify that the row is the same and to update the existing row instead. This is required because when new files are found they may need to be added to an existing row if they have the same angle and are in the same group as an existing run. In this case the row will then be re-processed and the runs preprocessed (summed together)

Re #17029