Allow runs to be added into a partial list
This commit allows new runs to be added into an existing row containing a partial list for that angle. This is required because during autoprocessing we do not have the full list of runs for a row up front. The first run is therefore processed but when the subsequent runs become available they should be added into the existing row rather than into separate rows.

We add a check that the existing list partially matches before adding the new row.
We consider the existing row to be an exact match if it already contains all runs in the new list (even if the new list is a partial set of the existing runs).

Re #17029