Add a Pause button to the runs tab
The new button does exactly the same as the pause button on the data processor. However, it is a much clearer to the user to have the pause button located immediately next to the Autoprocess button, so that they can easily see how to pause the autoprocessing they just started.

Re #17029