Don't start autoprocessing if user cancels
The user is given a warning if starting a new autoreduction will delete the current contents of the table. This commit fixes an issue where if they select to cancel the deletion, then the autoprocess was being run anyway, resulting in confusing behaviour where the new rows were appended to the existing rows.

Re #17029