Fixed bug with FindSubtreeRoots and ExtractSubtree.
Both algorithms would return non-empty optionals for tree selections
where a subtree with a root at greater depth came before a subtree
with a root at a lesser depth. Such as

[0, 0]
[0, 0, 0]
[0, 0, 1]
[1]
[1, 0]
[1, 1]
[1, 2]

This should not be the case as the expected behaviour for these
cases is ambiguous.

This commit fixes this behaviour for both algorithms and adds
unit tests for the above case.

Re #22263