Corrected input validation for group workspaces
Previously, only columns of m_inputWSMatrix (a vector of vector of shared matrix workspace pointers) were stitched either for matrix workspaces or groups of matrix workspaces.
Now, eiter a column or a row of m_inputWSMatrix will be stitched for matrix workspaces or groups of matrix workspaces, respectively.
This is only an implementation detail.

Needed to check the compatibility of each row of the m_inputWSMatrix

The code can be easier when using iterators ... for the vector of vector of shared matrix workspace pointers, but many times (probable) solution compiled well giving the error:
0x00007ffff50ef36d in boost::shared_ptr<Mantid::API::MatrixWorkspace>::shared_ptr (this=0x7720646e6966206f, r=...) at /usr/include/boost/smart_ptr/shared_ptr.hpp:392
392	    shared_ptr( shared_ptr const & r ) BOOST_NOEXCEPT : px( r.px ), pn( r.pn )

For instance, it would be nice to do m_inputWSMatrix[i].emplace_back(inputMatrix); or using iterators for each column and row of the matrix.

Please note as well that m_inputWSMatrix.reserve(inputWorkspacesStr.size()); reserves capacity well for group workspaces only, the worst case. For matrix workspaces not grouped it would be enough to m_inputWSMatrix.reserve(1);

Refs #22197