Move location of QApplication construction in mantidqt tests
Storing a reference to the QApplication object globally at
module scope causes the QApplication object to be destroyed
very late in Python's tear down process. This seems to cause
random segfaults with PyQt5 that can be remedied by using
object attributes so that the references are cleaned up much
sooner.