Set an error on rows when workspaces are removed
If a workspace is deleted or renamed in the ADS, then any rows for which it is a mandatory output workspace now have an error set on them rather than having their state reset. This avoids them being reprocessed automatically during autoprocessing. This avoids an infinite loop that was occurring if you have duplicate runs in the table because the second run will delete the output from the first, which will then delete the output from the second etc. and continue forever.

Re #17029