Fix a problem where users are not warned of errors
This commit removes a line of code which disables error messages when processing rows in the table. This means that a silent error occurs when users are manually processing rows, but the processing is stopped, so it looks like processing completed ok (especially if multiple rows were being processed). It looks like the code was a workaround and the original reason the code was added is no longer relevant. The logic to control whether a user receives prompts or not is (and should be) handled elsewhere (i.e. they are prompted if processing rows manually, but not if autoprocessing).

Re #17029