Don't mark workspace as deleted if still processing
If a workspace is deleted while a row is being reduced which will recreate it, then don't mark it as deleted. this fixes a bug where re-reducing a row which already has output causes the workspace to be deleted and recreated, and this was setting an error on the row incorrectly.

Re #17029