Refs #22158: This fixes `ResampleX` for `PreserveEvents=False`
Done by adding a `+1` to the number of bins in `ResampleX.cpp`.
In addition, tests actually executing the `ResampleX` algorithm were added to `ResampleXTest.h`.
These also illustrate the fact that the algorithm now works with `PreserveEvents=False`.
Some tests were also added to test the algorithm on a Workspace2D instead of just the EventWorkspace.
However, this only works if the workspace does not contain distribution data.
This issue has been reported in #22562.
Finally, a slightly more informative error message is thrown by the `VectorHelper::rebin` function.