Re #0 Hook up background thread to project recovery saving
Hooks up the background thread to project recovery saving. The thread
uses a conditional variable to avoid polling for changes. A mutex is
monitored and any changes (indicating thread exit) will cause the loop
to spin again. A timeout is then (ab)used to implement the period
between each save iteration

To save the GUI windows we must emit a signal back to the main window.
This is added to Qt's main event queue to execute. Workspace saving is
handled through a python script that is executed in a two line call.