Refs #22608 - Fix Windows warnings.
Included moving a TS_ASSERT out of a destructor in a mock object, which would have been harmless.  I couldn't use noexcept(false) as a workaround because of the parent destructor.  The resulting code is much messier as I have to keep ahold of the mock object and call the assert in each of the tests... :(