Re #22638 Remove chaining for ISIS run details on GEM
Removes chaining callbacks together to access a dictionary. This is one
of the things originally added to the scripts which should have been
removed.

This overly complex piece of code attempted to take the DRY pricnciple
to an extreme. Instead each instrument should just specify the layout of
their file