Fix bug caused by default instrument
The issue is that for some of the fields the instrument selected in the
GUI were ignored. This caused things to find the files for the default
instrument for some fields, and the GUI's instrument for others.

The actual change is a single line, but some modernization of the code
happened as well.