Re #22714 Fix various deadlocks from shutting down
Fixes various deadlocks in ProjectRecovery which can occur during
shutdown. These include holding a lock in the worker thread, then
calling into the GUI thread which is attempting to lock the mutex.

Waiting for the thread to join in the GUI thread, whilst the worker
thread is waiting for the GUI thread to save

Catching CancelExceptions and silently handling them