Re #22927 Fix binary operators not saving ws history
Fixes binary operators not saving ws history correctly. As we create a
workspace and use it's pointer there is no name associated. This means
that the pointer is suffixed onto `__tmp`.

When users go to load the history later the __tmp name doesn't match the
output property name used in CreateSingleValueWorkspace, so this fails.

Instead we register it into the ADS before executing the operator, then
remove it afterwards. This guarentees a name exists on the workspace as
a side effect of using the ADS