Fix unhandled exception in validateInputs
This commit adds a temporary fix for an unhandled exception in validateInputs. The exception occurs when opening AlgorithmDialog on a group workspace. This first calls validateInputs on the group for the dialog, before calling it on the individual workspaces when executing the algorithm. Workspace groups are not currently handled. Adding full validation for groups is beyond this PR so this fix just skips validation rather than throwing the exception.

Re #22933