Skip validation for workspace groups
This commit removes the validation error for workspace groups which results in an error shown on the GUI. Instead validation is skipped for group workspace inputs to the efficiencies workspace property. The reason for this is that validateInputs is not called when the algorithm is run from a script. Therefore we get the same behaviour from the gui and the script, where the algorithm will be called on each member of the group initially. This could be improved further but is out of scope of this PR.

Re #23061