Ignore whitespace when patching gtest
Ignore whitespace when patching gtest to avoid a failure in CMake. This
is because Mantid has a .gitattribute which forces auto CRLF, however
Google do not. So if the config option is set to false the clone will
have differing line endings causing the patch to fail without these
changes