Refs #20443 : Only using threaded Join() function if input is large.
If it is smaller than 500 * (Maximum number of threads), the method reverts to the original
unthreaded version of the function. This old function is also the one used if the input
array has random access iterators.

This should fix slow python.api tests and hopefully should get rid of the timeouts on build servers.