Refs #23285 : Re-arranged lists of tests and edited Jenkins files
Some tests were still failing because different modules are writing in the same directories, using the same files names.
There are at least 3 different locations the tests write to:
1. build/ExternalData/Testing/Data/SystemTest
2. build/Testing/SystemTests/scripts
3. build
In some cases a process tries to write to a file that already exists (same name used by another test).
In other cases a process performs a .cleanup() which removes a file that is still needed by a running test.
The solution was to group the tests in lists where all modules starting with the same 4 letters are given to one
core. This worsens the load balance between cores (with 8 cores, core 1 performs 93 tests while core 8 only has 44).
This is not ideal but allows the suite to complete without failures.
The runtime using 8 cores goes down from 2h to 30min, which is ok as a first step.