Remove reliance on peak ordering in WISH system test
The test required that a found peak was in a specific
position in the table but this is not guarnateed any longer
after bd4b4e3c88ec72e5565d4472a012395ef0d0a08f. The test failure
was not spotted on the original pull request as it was skipped
due to memory limitations on the node.