Centralise icon access in mantidqt and workbench
It was observed that a race condition between tests using
qtawesome could occur on macOS. A FontError would
be raised when trying to register new fonts but only when
the tests were run in parallel and never in isolation.

By centralising the access to qtawesome we can mock access
to it consistently. This also has the advantage of isolating
a thrid party API to help with future maintenance.