Refs #23373: Another test scheduling.
A master test list is first built from scanning the test python files.
For each module in this list, a list of data files required by that
module is created. An accompanying lock status for each data file is
also created.

Each thread then starts a loop and gathers a first test module from the
master test list which is stored in the tests_dict shared dictionary,
starting with the number in the module list equal to the process id.

Each process then checks if all the data files requird by the current
test module are available (i.e. have not been locked by another
thread). If all files are unlocked, the thread proceeds with that test
module. If not, it goes further down the list until it finds a module
whose files are all available.

Once it has completed the work in the current module, it checks if the
number of modules that remains to be executed is greater than 0. If
there is some work left to do, the thread finds the next module that
still has not been executed (searches through the tests_lock array
and finds the next element that has a 0 value).

This aims to have all threads end calculation approximately at the
same time, and also avoids having to modify many save directories
in the test files and the facilities interfaces they use.