Throw exception when there is unmatch column count
The version=1 file parsing assumed that there were more column labels
than values in a given row. This caused a segfault when reading past the
end of a vector. The code now checks for the values and labels to have
compatible lengths for each row and reports mismatch to the user. There
are also several changes in what particular exceptions are thrown
elsewhere in the algorithm.