Force progress messages to be str
The problem was that empty progress messages were being converted to
`None` which `AlgorithmObserver` did not have the proper overloaded
methods to convert to an empty string. The fix is to convert them to a
string before calling the binding. For extra measure, the progress
percentage is converted to a float as well.

Refs #23632