Reorganize ReflectometryILL
The calculation of several statistical quantities was extracted from
ReflectometryMomentumTransfer, updated to latest COSMOS convenstions,
and moved to a new algorithm, ReflectometryBeamStatistics.
This triggered a reorganization of ReflectometryILLSumForeground and
ReflectometryILLConvertToQ, the interface of the latter getting a
lot simpler thanks to the removal of the ReflectedBeamWorkspace and
DirectBeamWorkspace properties.

Re #23622