Reduce time to compute detector colors on instrument view
Introduces single-shot methods on the colormap types to compute
all colors from integrated counts in one go. For the matplotlib
implementation this avoids repeated expensive calls
to/from Python and recomputing of the normalization interval.
An example of computing the colors for an SXD dataset dropped
from 10s to ~0.1s in the matplotlib implementation.