Remove PythonRuntimeError exception type
Without specifying C++ exceptions that map to all of the
Python exceptions there is little benefit from only
distinguishing std:runtime_error form others. Now clients
only have to know that a single exception type could
be thrown. Swaps to std::runtime_error as base for
PythonException as its safer as Algorithm::exec doesn't
rethrow for std::runtime_error if it is not a child. I think this
is the behaviour we want.