Refs #23238 : tracked down the failed OSX build
This was due to the positions or rotation vectors created by the default
Eigen constructor. By default the components of the vecotrs are not initialised.
This can lead to different behaviours on different machines, if one compiler
zeros the values by default and another does not.
We are now making sure everything is initialised with a value in the tests.