Do not set empty event workspace to single bin
`EventWorkspace::resetAllXToSingleBin()` was getting nonsense values for
data range in an empty event workspace. The new behavior is to not
rebin the data in `LoadLiveData` to a single bin for this case. This
should move the segfault that was seen on REF_M to a `ValueError` the
first time `Rebin` is called in their post processing script.